Updated the test files consolidated under townhall instead of myapi
